Sidebar y reagrupación de algunos componentes.
