+++ src/App.css
@@ -1,3 +1,13 @@
+/* Global */
+
+.m-0 {
+  margin: 0;
+}
+
+.p-0 {
+  padding: 0;
+}
+
 .App {
   text-align: center;
 }
