+++ src/App.js
@@ -1,13 +1,11 @@
 import "./App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Menu from "./components/navbar/navbar";
 
-import AppRouter from "./components/router/router";
+import AppRouter from "./router/router";
 
 function App() {
   return (
     <div className="App">
-      <Menu />
       <AppRouter />
     </div>
   );
