--- src/components/pages/home/home.css
@@ -1,3 +1,0 @@
-h1 {
-  margin-top: 300px;
-}
