--- src/components/pages/home/home.js
@@ -1,12 +1,0 @@
-import React from "react";
-import "./home.css";
-
-export default class Home extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Ciclo Sport</h1>
-      </div>
-    );
-  }
-}
