+++ src/pages/dashboard/bicicletas/bicicletas.css
@@ -1,0 +1,25 @@
+#bicicletas-container {
+  margin-top: 50px;
+}
+
+#bicicletas-buscar-container,
+#bicicletas-crear-container {
+  margin-top: 50px;
+}
+
+.react-bs-table-sizePerPage-dropdown .dropdown-toggle {
+  float: left;
+  margin-bottom: 10px;
+}
+
+.search-label {
+  float: right;
+  padding-bottom: 10px;
+}
+.sr-only {
+  font-size: 0;
+}
+#bicicletas-crear-container .form-label {
+  float: left;
+  margin-bottom: 10px;
+}
