+++ src/pages/dashboard/empleados/crud/buscar.js
@@ -1,0 +1,163 @@
+import React from "react";
+import { request } from "../../../../components/helper/helper";
+import { Container, Row } from "react-bootstrap";
+
+import DataGrid from "../../../../components/grid/grid";
+import ConfirmationPromprs from "../../../../components/prompts/confirmation";
+import Loading from "../../../../components/loading/loading";
+import MessagePrompt from "../../../../components/prompts/message";
+
+const columns = [
+  {
+    dataField: "_id",
+    text: "Empleado ID",
+    hidden: true,
+  },
+  {
+    dataField: "nombre",
+    text: "Nombre",
+  },
+  {
+    dataField: "apellido_p",
+    text: "Apellido paterno",
+  },
+  {
+    dataField: "apellido_m",
+    text: "Apellido materno",
+  },
+  {
+    dataField: "telefono",
+    text: "Telefono",
+  },
+  {
+    dataField: "email",
+    text: "Email",
+  },
+  {
+    dataField: "direccion",
+    text: "Direccion",
+  },
+];
+
+export default class EmpleadosBuscar extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      idEmpleado: null,
+      confirmation: {
+        title: "Eliminar el empleado",
+        text: "¿Deseas eliminar el empleado?",
+        show: false,
+      },
+      message: {
+        text: "",
+        show: false,
+      },
+    };
+
+    this.onClickEditButton = this.onClickEditButton.bind(this);
+    this.onClickDeleteButton = this.onClickDeleteButton.bind(this);
+    this.onCancel = this.onCancel.bind(this);
+    this.onConfirm = this.onConfirm.bind(this);
+  }
+
+  onClickEditButton(row) {
+    this.props.setIdEmpleado(row._id);
+
+    this.props.changeTab("editar");
+  }
+
+  onClickDeleteButton(row) {
+    this.setState({
+      idEmpleado: row._id,
+      confirmation: {
+        ...this.state.confirmation,
+        show: true,
+      },
+    });
+  }
+
+  onCancel() {
+    this.setState({
+      confirmation: {
+        ...this.state.confirmation,
+        show: false,
+      },
+    });
+  }
+
+  onConfirm() {
+    this.setState(
+      {
+        confirmation: {
+          ...this.state.confirmation,
+          show: false,
+        },
+      },
+      this.eliminarEmpleado()
+    );
+  }
+
+  eliminarEmpleado() {
+    this.setState({ loading: true });
+    request
+      .delete(`/empleados/${this.state.idEmpleado}`)
+      .then((response) => {
+        this.setState({
+          loading: false,
+          message: {
+            text: response.data.msg,
+            show: true,
+          },
+        });
+        if (response.data.exito) this.reloadPage();
+      })
+      .catch((err) => {
+        console.error(err);
+        this.setState({ loading: false });
+      });
+  }
+
+  reloadPage() {
+    setTimeout(() => {
+      window.location.reload();
+    }, 2500);
+  }
+
+  render() {
+    return (
+      <Container id="empleados-buscar-container">
+        <ConfirmationPromprs
+          show={this.state.confirmation.show}
+          title={this.state.confirmation.title}
+          text={this.state.confirmation.text}
+          onCancel={this.onCancel}
+          onConfirm={this.onConfirm}
+        />
+
+        <MessagePrompt
+          text={this.state.message.text}
+          show={this.state.message.show}
+          duration={2500}
+          onExited={this.onExitedMessage}
+        />
+
+        <Loading show={this.state.loading} />
+
+        <Row>
+          <h1>Buscar empleados</h1>
+        </Row>
+        <Row>
+          <DataGrid
+            url="/empleados"
+            columns={columns}
+            showEditButton={true}
+            showDeleteButton={true}
+            onClickEditButton={this.onClickEditButton}
+            onClickDeleteButton={this.onClickDeleteButton}
+          />
+        </Row>
+      </Container>
+    );
+  }
+}
