+++ src/pages/dashboard/empleados/crud/crear.js
@@ -1,0 +1,134 @@
+import React from "react";
+import { Container, Form, Row, Button } from "react-bootstrap";
+import { request } from "../../../../components/helper/helper";
+import Loading from "../../../../components/loading/loading";
+import MessagePrompt from "../../../../components/prompts/message";
+
+export default class EmpleadosCrear extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      rediret: false,
+      message: {
+        text: "",
+        show: false,
+      },
+      loading: false,
+      empleado: {
+        nombre: "",
+        apellido_p: "",
+        apellido_m: "",
+        telefono: "",
+        email: "",
+        direccion: "",
+      },
+    };
+    this.onExitedMessage = this.onExitedMessage.bind(this);
+  }
+
+  setValue(index, value) {
+    this.setState({
+      empleado: {
+        ...this.state.empleado,
+        [index]: value,
+      },
+    });
+  }
+
+  guardarEmpleados() {
+    this.setState({ loading: true });
+    request
+      .post("/empleados", this.state.empleado)
+      .then((response) => {
+        if (response.data.exito) {
+          this.setState({
+            rediret: response.data.exito,
+            message: {
+              text: response.data.msg,
+              show: true,
+            },
+          });
+        }
+        this.setState({ loading: false });
+      })
+      .catch((err) => {
+        console.error(err);
+        this.setState({ loading: true });
+      });
+  }
+
+  onExitedMessage() {
+    if (this.state.rediret) this.props.changeTab("buscar");
+  }
+
+  render() {
+    return (
+      <Container id="empleados-crear-container">
+        <MessagePrompt
+          text={this.state.message.text}
+          show={this.state.message.show}
+          duration={2500}
+          onExited={this.onExitedMessage}
+        />
+
+        <Loading show={this.state.loading} />
+
+        <Row>
+          <h1>Crear Empleados</h1>
+        </Row>
+        <Row>
+          <Form>
+            <Form.Group className="mb-3" controlId="formBasic">
+              <Form.Label>Nombre</Form.Label>
+              <Form.Control
+                onChange={(e) => this.setValue("nombre", e.target.value)}
+              />
+            </Form.Group>
+
+            <Form.Group className="mb-3" controlId="formBasic">
+              <Form.Label>Apellido Paterno</Form.Label>
+              <Form.Control
+                onChange={(e) => this.setValue("apellido_p", e.target.value)}
+              />
+            </Form.Group>
+
+            <Form.Group className="mb-3" controlId="formBasic">
+              <Form.Label>Apellido Materno</Form.Label>
+              <Form.Control
+                onChange={(e) => this.setValue("apellido_m", e.target.value)}
+              />
+            </Form.Group>
+
+            <Form.Group className="mb-3" controlId="formBasic">
+              <Form.Label>Telefono</Form.Label>
+              <Form.Control
+                onChange={(e) => this.setValue("telefono", e.target.value)}
+              />
+            </Form.Group>
+
+            <Form.Group className="mb-3" controlId="formBasic">
+              <Form.Label>Email</Form.Label>
+              <Form.Control
+                onChange={(e) => this.setValue("email", e.target.value)}
+              />
+            </Form.Group>
+
+            <Form.Group className="mb-3" controlId="formBasic">
+              <Form.Label>Dirrecion</Form.Label>
+              <Form.Control
+                onChange={(e) => this.setValue("direccion", e.target.value)}
+              />
+            </Form.Group>
+
+            <Button
+              variant="primary"
+              onClick={() => console.log(this.guardarEmpleados())}
+            >
+              Guardar Empleado
+            </Button>
+          </Form>
+        </Row>
+      </Container>
+    );
+  }
+}
