+++ src/pages/dashboard/empleados/empleados.css
@@ -1,0 +1,25 @@
+#empleados-container {
+  margin-top: 50px;
+}
+
+#empleados-buscar-container,
+#empleados-crear-container {
+  margin-top: 50px;
+}
+
+.react-bs-table-sizePerPage-dropdown .dropdown-toggle {
+  float: left;
+  margin-bottom: 10px;
+}
+
+.search-label {
+  float: right;
+  padding-bottom: 10px;
+}
+.sr-only {
+  font-size: 0;
+}
+#empleados-crear-container .form-label {
+  float: left;
+  margin-bottom: 10px;
+}
