+++ src/pages/dashboard/repuestos/repuestos.css
